Refactor FaaS support in ClientMetadataHelper
Move FaasEnvironment enum and related static methods to a
top-level class, in preparation for use in DefaultServerMonitor.

JAVA-4936
